Streamlit and Gradio Workshop/analytics_dashboard_demo/requirements.txt
streamlit>=1.54
openai>=2.21
pandas>=2.2
plotly>=6.0
python-dotenv>=1.0
